寫完test了,但還沒寫Save&Load

+++ DrawingApp/PresentationModel/PresentationModel.cs
@@ -35,7 +35,7 @@ namespace DrawingApp.PresentationModel
         {
             for (int i = 0; i < _buttonStatus.Count; i++)
                 _buttonStatus[i] = true;
-            if (buttonTag != 3)
+            if (buttonTag != (int)3m)
             {
                 _buttonStatus[buttonTag] = false;
                 _model.IsSelected = false;
